Rectify "malformed URL" fix in sass file
A prior commit wrongly fixed a below transpiler error, by redeclaring
a slice in each iteration of a loop, instead of appending a value to it.
```
    Error: Error building site: TOCSS: failed to transform "sass/style.sass"
    (text/x-sass): "assets/sass/style.sass:4:9": malformed URL
```
The real problem was that the transpiler did notlike the commas in the url string
built in the loop:
https://fonts.googleapis.com/css?family=Noto+Sans:300,400,600,700

This commit reverts the assignment mistake fixes the error by quoting
the url inside the `url` function.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>